requirements.txt
altair==5.3.0
MiniSom==2.3.2
pillow==10.3.0
shap==0.45.0
streamlit==1.33.0
streamlit_javascript
streamlit_folium==0.20.0
tqdm==4.66.2
Unidecode==1.3.8
xgboost==2.0.3
st-pages==0.5.0
openpyxl
geopandas
matplotlib
mapclassify
plotly
kaleido
reportlab
pypdf